feat: New IHasAdditionalAttributes
