+++ templates/src/src/MyRazorLibrary/Components/Div.cs
@@ -4,8 +4,12 @@
 /// Represents a div HTML element.
 /// </summary>
 [HtmlTag("div")]
-public class Div : BlazorComponentBase, IHasChildContent
+public class Div : BlazorComponentBase, IHasChildContent, IHasAdditionalClass, IHasAdditionalStyle
 {
     /// <inheritdoc/>
     [Parameter]public RenderFragment? ChildContent { get; set; }
+    /// <inheritdoc/>
+    [Parameter]public string? AdditionalStyle { get; set; }
+    /// <inheritdoc/>
+    [Parameter]public string? AdditionalClass { get; set; }
 }
